Estimate A/B compartments from methylation array data

getArrayABsignal returns estimated A/B compartments from methylation array data.

Usage

getArrayABsignal(
  obj,
  res = 1000000,
  parallel = TRUE,
  chr = NULL,
  targets = NULL,
  preprocess = TRUE,
  cores = 2,
  bootstrap = TRUE,
  num.bootstraps = 1000,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  other = NULL,
  array.type = c("hm450", "EPIC"),
  group = FALSE,
  boot.parallel = TRUE,
  boot.cores = 2
)

Arguments

obj

Input SummarizedExperiment object

res

Compartment resolution in bp

parallel

Whether to run samples in parallel

chr

What chromosome to work on (leave as NULL to run on all chromosomes)

targets

Samples/cells to shrink towards

preprocess

Whether to preprocess the arrays prior to compartment inference

cores

How many cores to use when running samples in parallel

bootstrap

Whether we should perform bootstrapping of inferred compartments

num.bootstraps

How many bootstraps to run

genome

What genome to work on ("hg19", "hg38", "mm9", "mm10")

other

Another arbitrary genome to compute compartments on

array.type

What type of array is this ("hm450", "EPIC")

group

Whether to treat this as a group set of samples

boot.parallel

Whether to run the bootstrapping in parallel

boot.cores

How many cores to use for the bootstrapping

Value

A RaggedExperiment of inferred compartments

Examples

data("meth_array_450k_chr14", package = "compartmap")
array_compartments <- getArrayABsignal(array.data.chr14, parallel=FALSE, chr="chr14", bootstrap=FALSE, genome="hg19", array.type="hm450")
#> Loading required package: minfi
#> Warning: there is no package called ‘minfi’
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'seqinfo': unable to find required package ‘minfi’