Plots A/B compartment estimates on a per chromosome basis

Plot A/B compartments bins

plotAB(
  x,
  chr = NULL,
  what = c("score", "flip.score"),
  main = "",
  ylim = c(-1, 1),
  unitarize = FALSE,
  reverse = FALSE,
  top.col = "deeppink4",
  bot.col = "grey50",
  with.ci = FALSE,
  filter = TRUE,
  filter.min.eigen = 0.02,
  median.conf = FALSE
)

Arguments

x
The matrix obejct returned from getCompartments

chr
Chromosome to subset to for plotting

main
Title for the plot

ylim
Y-axis limits (default is -1 to 1)

unitarize
Should the data be unitarized? (not explicitly necessary for arrays)

reverse
Reverse the sign of the PC values?

top.col
Top (pos. PC values) chromatin color to be plotted

bot.col
Bottom (neg. PC values) chromatin color to be plotted

with.ci
Whether to plot confidence intervals

filter
Whether to filter eigenvalues close to zero (default: TRUE)

median.conf
Whether to plot the median confidence estimate across the chromosome

filter.lower.bound
The minimum absolute eigenvalue to include in the plot

Value

invisibly, the compartment estimates from the plot

Examples

library(GenomicRanges)
library(Homo.sapiens)

#Generate random genomic intervals of 1-1000 bp on chr1-22
#Modified from https://www.biostars.org/p/225520/
random_genomic_int <- data.frame(chr = rep("chr14", 100))
random_genomic_int$start <- apply(random_genomic_int, 1, function(x) { round(runif(1, 0, getSeqLengths(chr = x)[[1]]), 0) })

#> Error in getSeqLengths(chr = x): could not find function "getSeqLengths"

random_genomic_int$end <- random_genomic_int$start + runif(1, 1, 1000)

#> Error in `$<-.data.frame`(`*tmp*`, end, value = numeric(0)): replacement has 0 rows, data has 100

random_genomic_int$strand <- "*"

#Generate random counts
counts <- rnbinom(1000, 1.2, 0.4)

#Build random counts for 10 samples
count.mat <- matrix(sample(counts, nrow(random_genomic_int) * 10, replace = FALSE), ncol = 10)
colnames(count.mat) <- paste0("sample_", seq(1:10))

#Bin counts
bin.counts <- getBinMatrix(count.mat, makeGRangesFromDataFrame(random_genomic_int), chr = "chr14", genome = "hg19")

#> Error in .find_start_end_cols(df_colnames0, start.field0, end.field0): cannnot determine start/end columns

#Calculate correlations
bin.cor.counts <- getCorMatrix(bin.counts)

#> Calculating correlations...

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'cor': object 'bin.counts' not found

#Get A/B signal
absignal <- getABSignal(bin.cor.counts)

#> Calculating eigenvectors...

#> Error in t(matrix): object 'bin.cor.counts' not found

#Plot the A/B signal
par(mar=c(1,1,1,1))
par(mfrow=c(1,1))
plotAB(absignal$pc, ylim = c(-0.2, 0.2), unitarize = TRUE)

#> Error in is(x, "GRanges"): object 'absignal' not found

if (FALSE) {
#If plotting individual A/B signals using output from getCompartments
#Note: this function currently only supports plotting individual chromosomes from single samples
bin.chr1.ab <- getCompartments(data, "array", chrs = "chr1", genome = "hg19")

#For 7 samples
#Adjust ylim as necessary
par(mar=c(1,1,1,1))
par(mfrow=c(7,1))
plotAB(bin.chr1.ab[,1], ylim = c(-0.2, 0.2), unitarize = TRUE)
plotAB(bin.chr1.ab[,2], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "goldenrod")
plotAB(bin.chr1.ab[,3], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "darkblue")
plotAB(bin.chr1.ab[,4], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "red")
plotAB(bin.chr1.ab[,5], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "black")
plotAB(bin.chr1.ab[,6], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "cyan")
plotAB(bin.chr1.ab[,7], ylim = c(-0.2, 0.2), unitarize = TRUE, top.col = "seagreen")
}