Estimate A/B compartments from single-cell sequencing data

scCompartments returns estimated A/B compartments from sc-seq data.

Usage

scCompartments(
  obj,
  res = 1000000,
  parallel = FALSE,
  chr = NULL,
  targets = NULL,
  cores = 2,
  bootstrap = TRUE,
  num.bootstraps = 100,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  group = FALSE,
  assay = c("atac", "rna")
)

Arguments

obj

Input SummarizedExperiment object

res

Compartment resolution in bp

parallel

Whether to run samples in parallel

chr

What chromosome to work on (leave as NULL to run on all chromosomes)

targets

Samples/cells to shrink towards

cores

How many cores to use when running samples in parallel

bootstrap

Whether we should perform bootstrapping of inferred compartments

num.bootstraps

How many bootstraps to run

genome

What genome to work on ("hg19", "hg38", "mm9", "mm10")

group

Whether to treat this as a group set of samples

assay

What type of single-cell assay is the input data ("atac" or "rna")

Value

A RaggedExperiment of inferred compartments

Examples

data("k562_scrna_chr14", package = "compartmap")
sc_compartments <- scCompartments(
  k562_scrna_chr14,
  parallel = FALSE,
  chr = "chr14",
  bootstrap = FALSE,
  genome = "hg19"
)
#> Working on A10_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on A9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on B9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on C9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on D8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on E9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on F9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on G9_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H1_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H2_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H3_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H4_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H5_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H6_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H7_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.
#> Working on H8_hg19Aligned.sortedByCoord.out.bw
#> Computing compartments for chr14
#> Number of means fewer than 4. Using Bayes instead of JSE.
#> 108 bins created...
#> Calculating correlations...
#> Done...
#> Calculating eigenvectors.
#> Smoothing eigenvector.
#> Done smoothing.